Move administration page fixtures to dedicated fixtures file
